server: Factor out CompressionStream
